backend/requirements.txt
-c constraints.txt
-e src/haagshistorischmuseum[test]

zope.testrunner
kitconcept.seo
collective.volto.formsupport
redturtle.voltoplugin.editablefooter
collective.volto.socialsettings

# Add required add-ons
# (Ideally add them in setup.py for haagshistorischmuseum)
# collective.easyform
